RNG works. Need a seed or else generates same values.

+++ src/cudaFuncs.h
@@ -9,7 +9,7 @@
 #include <curand.h>
 void chol(double* , int , cublasFillMode_t );
 
-void normGen( double*, size_t, int, int);
+void normGen( double*, size_t, double, double);
 
 
 #endif
